Adding broadcast_to/broadcast_from to collectives module
- unify implementation of collectives
- factor out beginnings of communicator functionality
- fixing a couple of minor issues in existing collectives
- modernized API of gather_here/gather_there (futures are now taken by rvalue ref)
- renamed existing broadcast.hpp to broadcast_direct.hpp
