+++ libs/collectives/include/hpx/collectives/all_reduce.hpp
@@ -41,9 +41,9 @@ namespace hpx { namespace lcos {
     ///             usage of the \a HPX_REGISTER_ALLREDUCE macro to define the
     ///             necessary internal facilities used by \a all_reduce.
     ///
-    /// \returns    This function returns a future holding a vector with all
-    ///             values send by all participating sites. It will become
-    ///             ready once the all_reduce operation has been completed.
+    /// \returns    This function returns a future holding a value calculated
+    ///             based on the values send by all participating sites. It will
+    ///             become ready once the all_reduce operation has been completed.
     ///
     template <typename T, typename F>
     hpx::future<T> all_reduce(char const* basename, hpx::future<T> result,
@@ -136,215 +136,144 @@ namespace hpx { namespace lcos {
 
 #if !defined(HPX_COMPUTE_DEVICE_CODE)
 
-#include <hpx/assertion.hpp>
-#include <hpx/basic_execution/register_locks.hpp>
 #include <hpx/async/dataflow.hpp>
+#include <hpx/basic_execution.hpp>
+#include <hpx/collectives/detail/communicator.hpp>
 #include <hpx/functional/bind_back.hpp>
 #include <hpx/lcos/future.hpp>
-#include <hpx/local_lcos/and_gate.hpp>
-#include <hpx/synchronization/spinlock.hpp>
-#include <hpx/parallel/algorithms/reduce.hpp>
 #include <hpx/preprocessor/cat.hpp>
 #include <hpx/preprocessor/expand.hpp>
 #include <hpx/preprocessor/nargs.hpp>
 #include <hpx/runtime/basename_registration.hpp>
-#include <hpx/runtime/components/new.hpp>
-#include <hpx/runtime/components/server/component_base.hpp>
 #include <hpx/runtime/get_num_localities.hpp>
 #include <hpx/runtime/launch_policy.hpp>
 #include <hpx/runtime/naming/id_type.hpp>
-#include <hpx/runtime/naming/unmanaged.hpp>
+#include <hpx/traits/acquire_shared_state.hpp>
 #include <hpx/type_support/decay.hpp>
 #include <hpx/type_support/unused.hpp>
 
 #include <cstddef>
+#include <memory>
 #include <mutex>
 #include <string>
 #include <type_traits>
 #include <utility>
 #include <vector>
 
-namespace hpx { namespace lcos {
+namespace hpx { namespace traits {
 
-    namespace detail {
+    namespace communication {
+        struct all_reduce_tag;
+    }
 
-        ////////////////////////////////////////////////////////////////////////
-        template <typename T>
-        class all_reduce_server
-          : public hpx::components::component_base<all_reduce_server<T>>
+    ///////////////////////////////////////////////////////////////////////////
+    // support for all_reduce
+    template <typename Communicator>
+    struct communication_operation<Communicator, communication::all_reduce_tag>
+      : std::enable_shared_from_this<communication_operation<Communicator,
+            communication::all_reduce_tag>>
+    {
+        communication_operation(Communicator& comm)
+          : communicator_(comm)
         {
-            using mutex_type = lcos::local::spinlock;
+        }
 
-        public:
-            all_reduce_server()    //-V730
-            {
-                HPX_ASSERT(false);    // shouldn't ever be called
-            }
+        template <typename Result, typename T, typename F>
+        Result get(std::size_t which, T&& t, F&& op)
+        {
+            using arg_type = typename std::decay<T>::type;
+            using mutex_type = typename Communicator::mutex_type;
 
-            all_reduce_server(std::size_t num_sites, std::string const& name,
-                std::size_t site)
-              : num_sites_(num_sites)
-              , data_(num_sites)
-              , gate_(num_sites)
-              , name_(name)
-              , site_(site)
-            {
-            }
+            auto this_ = this->shared_from_this();
+            auto on_ready =
+                [this_ = std::move(this_), op = std::forward<F>(op)](
+                    hpx::shared_future<void> f) mutable -> arg_type {
+                f.get();    // propagate any exceptions
 
-            template <typename F>
-            hpx::future<T> get_result(std::size_t which, T t, F op)
-            {
-                std::unique_lock<mutex_type> l(mtx_);
+                auto& communicator = this_->communicator_;
 
-                auto on_ready = [this, op = std::move(op)](
-                                    hpx::shared_future<void> f) mutable -> T {
-                    f.get();    // propagate any exceptions
+                std::vector<arg_type> data;
+                {
+                    std::unique_lock<mutex_type> l(communicator.mtx_);
+                    data = communicator.data_;
+                }
 
-                    std::vector<T> data;
-                    std::string name;
-
-                    {
-                        std::unique_lock<mutex_type> l(mtx_);
-                        util::ignore_while_checking<
-                            std::unique_lock<mutex_type>>
-                            il(&l);
-                        data = data_;
-                        std::swap(name, name_);
-                    }
-
-                    // this is a one-shot object (generations counters are
-                    // not supported), unregister ourselves, but only once)
-                    if (!name.empty())
-                    {
-                        hpx::unregister_with_basename(name, site_).get();
-                    }
-
-                    HPX_ASSERT(!data.empty());
-
-                    auto it = data.begin();
-                    return hpx::parallel::reduce(hpx::parallel::execution::par,
-                        ++it, data.end(), *data.begin(), op);
-                };
-
-                hpx::future<T> f = gate_.get_shared_future(l).then(
-                    hpx::launch::async, std::move(on_ready));
-
-                gate_.synchronize(1, l);
-                data_[which] = std::move(t);
-                gate_.set(which, l);
-
-                return f;
-            }
-
-            template <typename F>
-            struct get_result_action
-              : hpx::actions::make_action<hpx::future<T> (all_reduce_server::*)(
-                                              std::size_t, T, F),
-                    &all_reduce_server::template get_result<F>,
-                    get_result_action<F>>::type
-            {
+                auto it = data.begin();
+                return hpx::parallel::reduce(hpx::parallel::execution::par,
+                    ++it, data.end(), *data.begin(), op);
             };
 
-        private:
-            mutex_type mtx_;
-            std::size_t num_sites_;
-            std::vector<T> data_;
-            lcos::local::and_gate gate_;
-            std::string name_;
-            std::size_t site_;
-        };
+            std::unique_lock<mutex_type> l(communicator_.mtx_);
+            util::ignore_while_checking<std::unique_lock<mutex_type>> il(&l);
 
-        ////////////////////////////////////////////////////////////////////////
-        inline hpx::future<hpx::id_type> register_all_reduce_name(
-            hpx::future<hpx::id_type>&& f, std::string basename,
-            std::size_t site)
-        {
-            hpx::id_type target = f.get();
+            hpx::future<arg_type> f =
+                communicator_.gate_.get_shared_future(l).then(
+                    hpx::launch::sync, std::move(on_ready));
 
-            // Register unmanaged id to avoid cyclic dependencies, unregister
-            // is done after all data has been collected in the component above.
-            hpx::future<bool> result = hpx::register_with_basename(
-                basename, hpx::unmanaged(target), site);
-
-            return result.then(hpx::launch::sync,
-                [target = std::move(target), basename = std::move(basename)](
-                    hpx::future<bool>&& f) -> hpx::id_type {
-                    bool result = f.get();
-                    if (!result)
-                    {
-                        HPX_THROW_EXCEPTION(bad_parameter,
-                            "hpx::lcos::detail::register_all_reduce_name",
-                            "the given base name for the all_reduce operation "
-                            "was already registered: " +
-                                basename);
-                    }
-                    return target;
-                });
+            communicator_.gate_.synchronize(1, l);
+            communicator_.data_[which] = std::forward<T>(t);
+            if (communicator_.gate_.set(which, l))
+            {
+                // this is a one-shot object (generations counters are not
+                // supported), unregister ourselves (but only once)
+                hpx::unregister_with_basename(
+                    std::move(communicator_.name_), communicator_.site_)
+                    .get();
+            }
+            return f;
         }
-    }    // namespace detail
 
-    ////////////////////////////////////////////////////////////////////////////
+        Communicator& communicator_;
+    };
+}}    // namespace hpx::traits
+
+namespace hpx { namespace lcos {
+
+    ///////////////////////////////////////////////////////////////////////////
     template <typename T>
     hpx::future<hpx::id_type> create_all_reduce(char const* basename,
         std::size_t num_sites = std::size_t(-1),
         std::size_t generation = std::size_t(-1),
         std::size_t this_site = std::size_t(-1))
     {
-        if (num_sites == std::size_t(-1))
-        {
-            num_sites = static_cast<std::size_t>(
-                hpx::get_num_localities(hpx::launch::sync));
-        }
-        if (this_site == std::size_t(-1))
-            this_site = static_cast<std::size_t>(hpx::get_locality_id());
-
-        std::string name(basename);
-        if (generation != std::size_t(-1))
-            name += std::to_string(generation) + "/";
-
-        // create a new all_reduce_server
-        using result_type = typename util::decay<T>::type;
-        hpx::future<hpx::id_type> id =
-            hpx::new_<detail::all_reduce_server<result_type>>(
-                hpx::find_here(), num_sites, name, this_site);
-
-        // register the all_reduce's id using the given basename
-        return id.then(hpx::launch::sync,
-            util::bind_back(
-                &detail::register_all_reduce_name, std::move(name), this_site));
+        return detail::create_communicator<T>(
+            basename, num_sites, generation, this_site);
     }
 
     ////////////////////////////////////////////////////////////////////////////
     // destination site needs to be handled differently
     template <typename T, typename F>
-    hpx::future<T> all_reduce(hpx::future<hpx::id_type>&& f,
+    hpx::future<T> all_reduce(hpx::future<hpx::id_type>&& fid,
         hpx::future<T>&& local_result, F&& op,
         std::size_t this_site = std::size_t(-1))
     {
         if (this_site == std::size_t(-1))
+        {
             this_site = static_cast<std::size_t>(hpx::get_locality_id());
+        }
 
         auto all_reduce_data =
             [op = std::forward<F>(op), this_site](hpx::future<hpx::id_type>&& f,
                 hpx::future<T>&& local_result) mutable -> hpx::future<T> {
             using func_type = typename std::decay<F>::type;
-            using action_type = typename detail::all_reduce_server<
-                T>::template get_result_action<func_type>;
+            using action_type = typename detail::communicator_server<T>::
+                template communication_get_action<
+                    traits::communication::all_reduce_tag, hpx::future<T>, T,
+                    func_type>;
 
             // make sure id is kept alive as long as the returned future
             hpx::id_type id = f.get();
             hpx::future<T> result = async(action_type(), id, this_site,
                 local_result.get(), std::forward<F>(op));
 
-            return result.then(hpx::launch::sync,
-                [id = std::move(id)](hpx::future<T>&& f) -> T {
-                    HPX_UNUSED(id);
-                    return f.get();
-                });
+            traits::detail::get_shared_state(result)->set_on_completed(
+                [id = std::move(id)]() { HPX_UNUSED(id); });
+
+            return result;
         };
 
         return dataflow(hpx::launch::sync, std::move(all_reduce_data),
-            std::move(f), std::move(local_result));
+            std::move(fid), std::move(local_result));
     }
 
     template <typename T, typename F>
@@ -360,9 +289,11 @@ namespace hpx { namespace lcos {
                 hpx::get_num_localities(hpx::launch::sync));
         }
         if (this_site == std::size_t(-1))
+        {
             this_site = static_cast<std::size_t>(hpx::get_locality_id());
+        }
 
-        if (this_site == 0)
+        if (this_site == root_site)
         {
             return all_reduce(create_all_reduce<T>(
                                   basename, num_sites, generation, root_site),
@@ -381,37 +312,39 @@ namespace hpx { namespace lcos {
     // all_reduce plain values
     template <typename T, typename F>
     hpx::future<typename std::decay<T>::type> all_reduce(
-        hpx::future<hpx::id_type>&& f, T&& local_result, F&& op,
+        hpx::future<hpx::id_type>&& fid, T&& local_result, F&& op,
         std::size_t this_site = std::size_t(-1))
     {
         if (this_site == std::size_t(-1))
+        {
             this_site = static_cast<std::size_t>(hpx::get_locality_id());
+        }
 
         using arg_type = typename std::decay<T>::type;
+
         auto all_reduce_data_direct =
             [op = std::forward<F>(op),
                 local_result = std::forward<T>(local_result),
                 this_site](hpx::future<hpx::id_type>&& f) mutable
             -> hpx::future<arg_type> {
             using func_type = typename std::decay<F>::type;
-            using action_type = typename detail::all_reduce_server<
-                arg_type>::template get_result_action<func_type>;
+            using action_type = typename detail::communicator_server<arg_type>::
+                template communication_get_action<
+                    traits::communication::all_reduce_tag,
+                    hpx::future<arg_type>, arg_type, func_type>;
 
             // make sure id is kept alive as long as the returned future
             hpx::id_type id = f.get();
             hpx::future<arg_type> result = async(action_type(), id, this_site,
                 std::forward<T>(local_result), std::forward<F>(op));
 
-            return result.then(hpx::launch::sync,
-                [id = std::move(id)](hpx::future<arg_type>&& f) -> arg_type {
-                    HPX_UNUSED(id);
-                    return f.get();
-                });
+            traits::detail::get_shared_state(result)->set_on_completed(
+                [id = std::move(id)]() { HPX_UNUSED(id); });
+
+            return result;
         };
 
-        using result_type = typename util::decay<T>::type;
-        return dataflow(
-            hpx::launch::sync, all_reduce_data_direct, std::move(f));
+        return fid.then(hpx::launch::sync, std::move(all_reduce_data_direct));
     }
 
     template <typename T, typename F>
@@ -426,7 +359,9 @@ namespace hpx { namespace lcos {
                 hpx::get_num_localities(hpx::launch::sync));
         }
         if (this_site == std::size_t(-1))
+        {
             this_site = static_cast<std::size_t>(hpx::get_locality_id());
+        }
 
         if (this_site == root_site)
         {
@@ -469,7 +404,7 @@ namespace hpx {
 
 #define HPX_REGISTER_ALLREDUCE_2(type, name)                                   \
     typedef hpx::components::component<                                        \
-        hpx::lcos::detail::all_reduce_server<type>>                            \
+        hpx::lcos::detail::communicator_server<type>>                          \
         HPX_PP_CAT(all_reduce_, name);                                         \
     HPX_REGISTER_COMPONENT(HPX_PP_CAT(all_reduce_, name))                      \
     /**/
