+++ libs/collectives/include/hpx/collectives/spmd_block.hpp
@@ -7,7 +7,7 @@
 #pragma once
 
 #include <hpx/collectives/barrier.hpp>
-#include <hpx/collectives/broadcast.hpp>
+#include <hpx/collectives/broadcast_direct.hpp>
 #include <hpx/concepts/concepts.hpp>
 #include <hpx/execution/execution.hpp>
 #include <hpx/functional/first_argument.hpp>
