+++ libs/collectives/include_compatibility/hpx/lcos/broadcast.hpp
@@ -8,14 +8,14 @@
 
 #include <hpx/config.hpp>
 #include <hpx/collectives/config/defines.hpp>
-#include <hpx/collectives/broadcast.hpp>
+#include <hpx/collectives/broadcast_direct.hpp>
 
 #if defined(HPX_COLLECTIVES_HAVE_DEPRECATION_WARNINGS)
 #if defined(HPX_MSVC)
 #pragma message("The header hpx/lcos/broadcast.hpp is deprecated, \
-    please include hpx/collectives/broadcast.hpp instead")
+    please include hpx/collectives/broadcast_direct.hpp instead")
 #else
 #warning "The header hpx/lcos/broadcast.hpp is deprecated, \
-    please include hpx/collectives/broadcast.hpp instead"
+    please include hpx/collectives/broadcast_direct.hpp instead"
 #endif
 #endif
